+++ tests/ExAs.Tests/Api/Enumerables/Enumerable_HasNone_WithPredicate_Feature.cs
@@ -1,8 +1,10 @@
-﻿using ExAs.Results;
+﻿using System;
+using System.Linq.Expressions;
+using ExAs.Results;
+using ExAs.Utils;
 using NUnit.Framework;
 using static ExAs.Utils.ComposeLog;
 using static ExAs.Utils.Creation.CreateNinjas;
-using static System.Environment;
 
 namespace ExAs.Api.Enumerables
 {
@@ -18,7 +20,7 @@ namespace ExAs.Api.Enumerables
             // assert
             result.ExAssert(r => r.Member(x => x.succeeded).IsTrue()
                                   .Member(x => x.actual)    .IsEqualTo("WeaponedNinja: ( )Weapons = <empty>")
-                                  .Member(x => x.expectation).IsEqualTo(Expected("0 matches")));
+                                  .Member(x => x.expectation).IsEqualTo(Expected("none fulfils w => w.StartsWith(\"Sw\")")));
         }
 
         [Test]
@@ -28,12 +30,31 @@ namespace ExAs.Api.Enumerables
             Result result = NinjaWithTwoWeapons().Evaluate(n => n.Member(x => x.Weapons).HasNone(w => w.StartsWith("Shu")));
 
             // assert
+            Console.WriteLine(result.PrintLog());
+            result.ExAssert(r => r.Member(x => x.succeeded) .IsFalse()
+                                  .Member(x => x.PrintLog()).IsEqualTo($"WeaponedNinja: (X)Weapons = ['Sword', 'Shuriken'] (expected: none fulfils w => w.StartsWith(\"Shu\"))"));
+        }
+
+        [Test]
+        public void ExpectingNullPredicate_OnWeaponlessNinja_ShouldFail()
+        {
+            // act
+            Result result = WeaponlessNinja().Evaluate(n => n.Member(x => x.Weapons).HasNone((Expression<Func<string, bool>>)null));
+
+            // assert
             result.ExAssert(r => r.Member(x => x.succeeded)  .IsFalse()
-                                  .Member(x => x.actual)     .IsEqualTo($"WeaponedNinja: (X)Weapons = <1 match>")//{NewLine}" +"")
-                                                                        //$"                            'Sword'{NewLine}" +
-                                                                        //$"                            'Shuriken'")
-                                  .Member(x => x.expectation).IsEqualTo($"1 match{NewLine}" +
-                                                                        $""));
+                                  .Member(x => x.expectation).IsEqualTo(Expected("none fulfils null")));
+        }
+
+        [Test]
+        public void ExpectingNoneStartingWithSw_OnNullWeapons_ShouldSucceed()
+        {
+            // act
+            Result result = WeaponlessNinja().Evaluate(n => n.Member(x => x.NullWeaponList()).HasNone(w => w.StartsWith("Sw")));
+
+            // assert
+            result.ExAssert(r => r.Member(x => x.succeeded) .IsTrue()
+                                  .Member(x => x.PrintLog()).IsEqualTo("WeaponedNinja: ( )NullWeaponList() = null (expected: none fulfils w => w.StartsWith(\"Sw\"))"));
         }
     }
 }